refactor(search): move search component to outside folders
