+++ components/series/SeriesTable.vue
@@ -344,7 +344,7 @@
 import { Component, Watch, mixins } from 'nuxt-property-decorator'
 
 import { Collection, NFTMetadata } from '@/components/rmrk/service/scheme'
-import { exist } from '@/components/rmrk/Gallery/Search/exist'
+import { exist } from '@/components/search/exist'
 import { sanitizeIpfsUrl } from '@/components/rmrk/utils'
 
 import AuthMixin from '@/utils/mixins/authMixin'
